+++ src/components/Button/index.ts
@@ -1,1 +1,2 @@
 export { default } from './Button'
+export { default as ButtonGroup } from './ButtonGroup'